Fin du cours

+++ php/BDDRequestPanier.php
@@ -12,25 +12,30 @@ if ($_GET['mode'] == 'ajouter') {
 } else if ($_GET['mode'] == 'afficher') {
     $panierRequest = "SELECT * FROM panier";
     $panierResponse = $mysqli->query($panierRequest);
+    if ($panierResponse->rowCount() == 0) {
+        echo("<h3>Votre panier est vide</h3>");
+    } else {
+        echo("<h3>Votre panier</h3>");
 
-    while ($rowpanier = $panierResponse -> fetch_object()) {
-        $presentationRequest = "SELECT * FROM presentationJeux WHERE id=" .$rowpanier->idJeux . " ";
-        $responsePresReq = $mysqli->query($presentationRequest);
+        while ($rowpanier = $panierResponse -> fetch_object()) {
+                $presentationRequest = "SELECT * FROM presentationJeux WHERE id=" .$rowpanier->idJeux . " ";
+                $responsePresReq = $mysqli->query($presentationRequest);
 
-    
-        $rowPresentation = $responsePresReq -> fetch_object();
-        if (strlen($rowPresentation->presentationJeu) > 100) {
-            $presentationJeu = substr($rowPresentation->presentationJeu,0,100) . "...";
-        }
-        echo ("
-                <section class=\"flexBoxColumn flexBoxCenter shadow white\">
-                    <h3 class=\"flexBoxRow flexBoxCenter\">".$rowPresentation->nomJeu."</h3>
-                    <a class=\"imageContainer\"  href=\"description.php?id=".$rowPresentation->id."\">
-                        <img src=\"./img/".$rowPresentation->nomFichier.".png\" alt=\"$rowPresentation->nomJeu\"/>
-                    </a>
-                    <figurecaption class=\"widthFull flexBoxRow flexBoxCenter\">".$presentationJeu."</figurecaption>  
-                    <a class=\"margin widthAuto bouton flexBoxRow flexBoxCenter shadow\" href=\"./php/BDDRequestPanier.php?mode=supprimer&id=".$rowpanier->idJeux."\">Supprimer du panier</a>     
-                </section>");
+            
+                $rowPresentation = $responsePresReq -> fetch_object();
+                if (strlen($rowPresentation->presentationJeu) > 100) {
+                    $presentationJeu = substr($rowPresentation->presentationJeu,0,100) . "...";
+                }
+                echo ("
+                        <section class=\"flexBoxColumn flexBoxCenter shadow white\">
+                            <h3 class=\"flexBoxRow flexBoxCenter\">".$rowPresentation->nomJeu."</h3>
+                            <a class=\"imageContainer\"  href=\"description.php?id=".$rowPresentation->id."\">
+                                <img src=\"./img/".$rowPresentation->nomFichier.".png\" alt=\"$rowPresentation->nomJeu\"/>
+                            </a>
+                            <figurecaption class=\"widthFull flexBoxRow flexBoxCenter\">".$presentationJeu."</figurecaption>  
+                            <a class=\"margin widthAuto bouton flexBoxRow flexBoxCenter shadow\" href=\"./php/BDDRequestPanier.php?mode=supprimer&id=".$rowpanier->idJeux."\">Supprimer du panier</a>     
+                        </section>");
+        } 
     }
 } else if ($_GET['mode'] == 'supprimer') {    
     $mysqli->query("DELETE FROM panier WHERE idJeux = " . $_GET['id'] . "");
